requirements.txt
advent-of-code-data
beautifulsoup4
black
certifi
charset-normalizer
click
constantly
dateutils
decorator
docopt
Flask
gitdb
GitPython
idna
incremental
itsdangerous
Jinja2
MarkupSafe
more-itertools
mpmath
mypy
mypy-extensions
networkx
numpy
packaging
pathspec
Pebble
Pillow
platformdirs
Protego
pyasn1
pyasn1-modules
pycparser
PyDispatcher
pyparsing
pypi
python-dateutil
pytz
pytz-deprecation-shim
queuelib
regex
requests
semantic-version
six
smmap
soupsieve
sympy
termcolor
typing_extensions
tzdata
tzlocal
urllib3
w3lib
Werkzeug
z3-solver
zope.interface
